requirements.txt

streamlit
pdfplumber
python-docx
langdetect
spacy
nltk
fpdf
openai
